Change the type of tpe field from RWST to State
tpe field in Tree trait had the type of some RWST, but this is an
overkill and we don't need anything beyond its state features, therefore
it was changed to State monad.

If there is an easy way to switch between/or interoperate reader-state
monad, then it is even better if it is changed to a Reader monad.