Introduce nameAtParser for use trees.
Sana takes a somewhat unique approach for name resolution. It has three
categories of trees:

1. Def Trees (or better Identified Trees), the trees that introduce a new name
2. Use Trees, the class of trees that use the ones defined by Def trees.
3. Other trees, that do not use or define any name.

Each Def Tree in Sana is assigned a unique id during the **IDAssigner** phase,
At the **AssignOwners** phase the compiler assign contextual owners for every
trees. Then the **Namers** phase comes and resolves the Uses to the Defs. How
Sana knows which id belongs to which Def? Simply by storing them in some
*Dictionary*, and thanks to the *State Monad* we can have this without giving
up immutability. For more information please see `tiny.util.TreeContexts` and
`tiny.util.MonadUtils`.

This approach helps us to easily change any property of a Def tree, and still
see the change in all its uses, without any additional effort. This is
particularly true for renaming a Def tree.

There is a subtle problem, at the parser phase we need to store the name of a
use, so at the **Namers** we know what was the name of this use at the source
code level. But after **Namers** we don't want this field to be used anymore
and instead we need to lookup the *Dictionary* for the `id` that this Use tree
uses.

The way we do it now is by having two different members in TypeUse, one is a
field `nameAtParser` which optionally returns the name of the tree when parsed.
The other is `name`, which returns a state monad that computes the name of the
tree that this tree uses by following its `id`.

`name` is a state monad that computes `Name`, which is a value class of one
single String member, and `nameAtParser` is of type `Option[String]` and that
makes the confusion a bit less (hopefully). And we intentionally do not provide
any implicit conversion between `String` and `Name`.

For more refined (and upto date) information refer to Name Resolution page in
the wiki.