Combine Namers and Name-Before-Use checker phases
These two phases are actually so close to each other. Dividing them into
two different phases just makes things harder.

For name resolution, if we don't have a clue about the order of
definition, we may simply resolve an ident to a name that will be
defined later in the scope, because the context has no notion of
geographical location of a tree.